# Fix the filtering for external tables in the Redshift get_columns_in_relation macro